Add super admin feature with artisan commands

+++ app/Http/Controllers/ProfileController.php
@@ -85,6 +85,13 @@ class ProfileController extends Controller
         
         $user = Auth::user();
         
+        // Prevent deletion of super admin users
+        if ($user->is_super_admin) {
+            return back()->withErrors([
+                'delete' => 'Super admin accounts cannot be deleted through this interface.'
+            ]);
+        }
+        
         Auth::logout();
         
         $user->delete();
